requirements.txt
streamlit
spacy
pdfminer.six
textblob
matplotlib
pillow
numpy
